refactor(tuiv2): Phase 10 Step 8 — 拆分 app.go 回到 <800 行约束
app.go 内联路由/命令行/搜索后膨胀至 1405 行，违反 AGENTS.md 单文件
<800 行约束，按职责拆分为同包多文件：

- app.go (785行)：保留 App 结构、Init/Update/View、handleKeyMsg/
  handleInputModeKey/handleCtrlC、toggle/trigger 辅助、appendStream/
  bindComponents 等核心装配
- app_normal.go (215行)：handleNormalModeKey + Ex/Search/n/N 辅助
  (enterInputFromNormal/openEx/openSearch/clearSearchAndEx/
  executeExCommand/executeSearch/jumpSearchMatch/scrollToStreamIndex)
- app_leader.go (132行)：handleLeaderKey + cancelCurrentRun/
  retryLastRun/switchToLastSession
- app_commands.go (254行)：Gateway 客户端 tea.Cmd 工厂与消息类型
  (loadInitialCmd/submitMessageCmd/cancelRunCmd/loadSessionCmd/
  createSessionCmd 等 + initialLoadedMsg/gatewayEventMsg 等)
- app_view.go (147行)：视图布局与尺寸适配
  (applyWindowSize/mainArea/separatorLine/fitViewToTerminal/fitLine/
  debugLine/applyInitialLoaded)

纯代码移动，无逻辑变更，全量测试通过。
