+++ internal/tuiv2/app.go
@@ -8,7 +8,6 @@ import (
 	"time"
 
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
 
 	"neo-code/internal/tuiv2/components"
 	"neo-code/internal/tuiv2/gateway"
@@ -251,24 +250,6 @@ func (a *App) View() string {
 	return a.fitViewToTerminal(strings.Join(lines, "\n"))
 }
 
-// applyWindowSize 更新布局尺寸，并按 Focus-Only 断点计算 Soft Inspector 状态。
-func (a *App) applyWindowSize(width int, height int) {
-	a.state.Layout.Width = width
-	a.state.Layout.Height = height
-	switch {
-	case width < inspectorHiddenWidth:
-		a.state.Layout.ShowInspector = false
-		a.state.Layout.InspectorWidth = 0
-	case width < inspectorWideMin:
-		a.state.Layout.ShowInspector = true
-		a.state.Layout.InspectorWidth = width
-	default:
-		a.state.Layout.ShowInspector = true
-		a.state.Layout.InspectorWidth = inspectorWideWidth
-	}
-}
-
-// routeComponents 将全局消息转发给各静态布局组件。
 func (a *App) routeComponents(msg tea.Msg) tea.Cmd {
 	_, statusCmd := a.ambientStatus.Update(msg)
 	_, streamCmd := a.agentStream.Update(msg)
@@ -401,164 +382,6 @@ func (a *App) handleInputModeKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
 	}
 }
 
-// handleNormalModeKey 处理 Normal Mode 下的键盘输入。
-//
-// 分层约定（plan-v4）：模式切换键（i/Enter→Input、Space→Leader、:→Ex、
-// /→Search）优先拦截；n/N 在搜索 Matches 非空时跳转；Ctrl+D 半页下翻；
-// 其余导航键交给 stream。
-func (a *App) handleNormalModeKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	action := keymap.MatchNormalKey(msg.String())
-	switch action {
-	case keymap.ActionCtrlC:
-		return a.handleCtrlC()
-	case keymap.ActionEnterInput:
-		a.enterInputFromNormal()
-		return a, nil
-	case keymap.ActionScrollDown, keymap.ActionScrollUp,
-		keymap.ActionScrollTop, keymap.ActionScrollBottom:
-		_, cmd := a.agentStream.Update(msg)
-		return a, cmd
-	case keymap.ActionHalfPageDown, keymap.ActionHalfPageUp:
-		_, cmd := a.agentStream.Update(msg)
-		return a, cmd
-	case keymap.ActionFullPageDown, keymap.ActionFullPageUp:
-		_, cmd := a.agentStream.Update(msg)
-		return a, cmd
-	case keymap.ActionLeader:
-		a.state.Mode = state.LeaderMode
-		return a, leaderTimeoutCmd()
-	case keymap.ActionQuit:
-		return a, tea.Quit
-	case keymap.ActionSearchForward:
-		// / 进入搜索输入 overlay
-		a.openSearch()
-		return a, nil
-	case keymap.ActionSearchNext:
-		a.jumpSearchMatch(1)
-		return a, nil
-	case keymap.ActionSearchPrev:
-		a.jumpSearchMatch(-1)
-		return a, nil
-	case keymap.ActionExCommand:
-		// : 进入 Ex 命令行输入 overlay
-		a.openEx()
-		return a, nil
-	default:
-		_, promptCmd := a.commandPrompt.Update(msg)
-		return a, promptCmd
-	}
-}
-
-// handleLeaderKey 处理 Leader Key 后缀。
-//
-// 行为约定（plan-v4）：Leader 是独占捕获，非后缀键或超时(1s)时立即静默回到
-// Normal（不泄漏给 Normal handler）。后缀键执行动作后回到 Normal，除非打开了
-// 需要保持的面板（palette/session_picker/help/model_picker）。
-func (a *App) handleLeaderKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	keyStr := msg.String()
-	if keyStr == "esc" || keyStr == "ctrl+c" {
-		a.state.Mode = state.NormalMode
-		return a, nil
-	}
-	action := keymap.MatchLeaderKey(keyStr)
-	// 非后缀键：静默回到 Normal，不执行任何动作。
-	if action == keymap.ActionNone {
-		a.state.Mode = state.NormalMode
-		return a, nil
-	}
-	a.state.Mode = state.NormalMode // Leader 后总是回到 Normal
-	switch action {
-	case keymap.ActionLeaderQuit:
-		return a, tea.Quit
-	case keymap.ActionLeaderPalette:
-		a.openOverlay(state.OverlayPalette)
-		return a, nil
-	case keymap.ActionLeaderHelp:
-		a.openOverlay(state.OverlayHelp)
-		return a, nil
-	case keymap.ActionLeaderSwitchSession:
-		a.openOverlay(state.OverlaySessionPicker)
-		return a, nil
-	case keymap.ActionLeaderModelPicker:
-		a.openOverlay(state.OverlayModelPicker)
-		return a, nil
-	case keymap.ActionLeaderNewSession:
-		if a.client != nil {
-			return a, createSessionCmd(a.client)
-		}
-		return a, nil
-	case keymap.ActionLeaderFullAccess:
-		return a, a.toggleFullAccess()
-	case keymap.ActionLeaderLog:
-		a.appendStream(state.StreamEntry{
-			ID:        fmt.Sprintf("log-hint-%d", time.Now().UnixNano()),
-			Type:      "status",
-			Timestamp: time.Now(),
-			Content:   "Log viewer not yet available",
-			Metadata:  map[string]any{"done": true},
-		})
-		return a, nil
-	case keymap.ActionLeaderCancelRun:
-		return a, a.cancelCurrentRun()
-	case keymap.ActionLeaderRetry:
-		return a, a.retryLastRun()
-	case keymap.ActionLeaderLastSession:
-		return a, a.switchToLastSession()
-	default:
-		return a, nil
-	}
-}
-
-// cancelCurrentRun 取消当前运行中的 Agent；无运行中任务时静默 no-op。
-func (a *App) cancelCurrentRun() tea.Cmd {
-	phase := a.state.Runtime.Phase
-	if phase != state.RuntimePhaseRunning &&
-		phase != state.RuntimePhaseWaitingPermission &&
-		phase != state.RuntimePhaseWaitingUser {
-		// 空闲态：静默 no-op，避免打扰用户。
-		return nil
-	}
-	if a.client != nil {
-		return cancelRunCmd(a.client, a.activeSessionID(), a.state.Runtime.RunID)
-	}
-	a.state.Runtime.Phase = state.RuntimePhaseCancelled
-	return nil
-}
-
-// retryLastRun 重试最近一次用户输入；无历史输入时提示。
-func (a *App) retryLastRun() tea.Cmd {
-	if strings.TrimSpace(a.lastUserText) == "" {
-		a.appendStream(state.StreamEntry{
-			ID:        fmt.Sprintf("retry-hint-%d", time.Now().UnixNano()),
-			Type:      "status",
-			Timestamp: time.Now(),
-			Content:   "No previous run to retry",
-			Metadata:  map[string]any{"done": true},
-		})
-		return nil
-	}
-	return a.handleSubmitMessage(components.SubmitMessageMsg{Text: a.lastUserText})
-}
-
-// switchToLastSession 切换到上一个会话；无上一会话时提示。
-func (a *App) switchToLastSession() tea.Cmd {
-	if a.prevSessionID == "" {
-		a.appendStream(state.StreamEntry{
-			ID:        fmt.Sprintf("last-sess-hint-%d", time.Now().UnixNano()),
-			Type:      "status",
-			Timestamp: time.Now(),
-			Content:   "No previous session to switch",
-			Metadata:  map[string]any{"done": true},
-		})
-		return nil
-	}
-	if a.client == nil {
-		return nil
-	}
-	return loadSessionCmd(a.client, a.prevSessionID)
-}
-
-// handleCtrlC 实现 Ctrl+C 双退保护：运行中取消、空闲双退。
 func (a *App) handleCtrlC() (tea.Model, tea.Cmd) {
 	phase := a.state.Runtime.Phase
 	if phase == state.RuntimePhaseRunning || phase == state.RuntimePhaseWaitingPermission || phase == state.RuntimePhaseWaitingUser {
@@ -675,127 +498,6 @@ func (a *App) closeOverlay() {
 	a.state.Overlay.Selected = 0
 }
 
-// enterInputFromNormal 从 Normal 进入 Input Mode，并清除 Normal 专属子状态（搜索）。
-func (a *App) enterInputFromNormal() {
-	a.state.Mode = state.InputModeInput
-	a.clearSearchAndEx()
-}
-
-// openEx 打开 : 命令行输入 overlay。
-func (a *App) openEx() {
-	a.state.Ex.Active = true
-	a.state.Ex.Input = ""
-	a.openOverlay(state.OverlayEx)
-}
-
-// openSearch 打开 / 搜索输入 overlay。
-func (a *App) openSearch() {
-	a.state.Search.Active = true
-	a.state.Search.Query = ""
-	a.openOverlay(state.OverlaySearch)
-}
-
-// clearSearchAndEx 清除搜索与 Ex 输入状态（切出 Normal 或事件触发时调用）。
-func (a *App) clearSearchAndEx() {
-	a.state.Search = state.SearchState{}
-	a.state.Ex = state.ExState{}
-}
-
-// executeExCommand 解释并执行 : 命令（已去除前缀 ":"），返回副作用 cmd。
-//
-// 支持命令：q/quit/exit=退出、debug=切调试、help=开帮助、compact=触发压缩、
-// mode=切换 Agent 模式。空或未知命令给出提示。
-func (a *App) executeExCommand(command string) tea.Cmd {
-	switch command {
-	case "q", "quit", "exit":
-		return tea.Quit
-	case "debug":
-		a.debug = !a.debug
-		a.appendStream(state.StreamEntry{
-			ID:        fmt.Sprintf("debug-toggle-%d", time.Now().UnixNano()),
-			Type:      "status",
-			Timestamp: time.Now(),
-			Content:   fmt.Sprintf("Debug: %v", a.debug),
-			Metadata:  map[string]any{"done": true},
-		})
-		return nil
-	case "help":
-		a.openOverlay(state.OverlayHelp)
-		return nil
-	case "compact":
-		return a.triggerCompact()
-	case "mode":
-		return a.toggleAgentMode()
-	default:
-		a.appendStream(state.StreamEntry{
-			ID:        fmt.Sprintf("ex-unknown-%d", time.Now().UnixNano()),
-			Type:      "status",
-			Timestamp: time.Now(),
-			Content:   fmt.Sprintf("Unknown ex command: %s", emptyDash(command)),
-			Metadata:  map[string]any{"done": true},
-		})
-		return nil
-	}
-}
-
-// executeSearch 执行全量扫描并记录匹配索引到 Search.Matches，滚动到首个匹配。
-//
-// 空 query 为 no-op（关闭搜索 overlay）；无匹配给出提示。
-func (a *App) executeSearch(query string) tea.Cmd {
-	if strings.TrimSpace(query) == "" {
-		return nil
-	}
-	matches := components.RunSearch(a.state.Stream, query)
-	a.state.Search.Matches = matches
-	a.state.Search.MatchIndex = 0
-	a.state.Search.Stale = false
-	if len(matches) == 0 {
-		a.appendStream(state.StreamEntry{
-			ID:        fmt.Sprintf("search-empty-%d", time.Now().UnixNano()),
-			Type:      "status",
-			Timestamp: time.Now(),
-			Content:   fmt.Sprintf("No matches: %s", query),
-			Metadata:  map[string]any{"done": true},
-		})
-		return nil
-	}
-	a.scrollToStreamIndex(matches[0])
-	return nil
-}
-
-// jumpSearchMatch 在搜索匹配间循环跳转（direction=1 下一个，-1 上一个）。
-//
-// 无匹配时静默 no-op；到末尾/首位循环折返。
-func (a *App) jumpSearchMatch(direction int) {
-	matches := a.state.Search.Matches
-	if len(matches) == 0 {
-		return
-	}
-	a.state.Search.MatchIndex = (a.state.Search.MatchIndex + direction + len(matches)) % len(matches)
-	a.scrollToStreamIndex(matches[a.state.Search.MatchIndex])
-}
-
-// scrollToStreamIndex 滚动 stream 使指定全局 entry 索引尽量可见。
-//
-// 由于 state.Stream 是 append-only 且全量在内存，这里基于目标索引估算
-// 滚动偏移（粗略：将目标定位到视口中部），足够满足跳转可见需求。
-func (a *App) scrollToStreamIndex(targetIndex int) {
-	if targetIndex < 0 || targetIndex >= len(a.state.Stream) {
-		return
-	}
-	// 粗略估计：stream 行数约为 entry 数的倍数，这里直接用 entry 索引作为
-	// 偏移参考，关闭自动滚动并尝试把目标带到视口。精确视口定位由 stream
-	// 渲染时的 visibleLines 兜底（超出范围会被 clamp）。
-	a.state.Layout.AutoScroll = false
-	// 反向估算：偏移越大表示越靠顶部。目标越靠后(索引大)越接近底部，偏移越小。
-	estimated := len(a.state.Stream) - targetIndex
-	if estimated < 0 {
-		estimated = 0
-	}
-	a.state.Layout.ScrollOffset = estimated
-}
-
-// handlePaletteCommand 处理命令面板选择的命令。
 func (a *App) handlePaletteCommand(msg components.PaletteCommandMsg) tea.Cmd {
 	switch msg.Name {
 	case "/exit":
@@ -995,97 +697,6 @@ func (a *App) activeSessionTitle() string {
 	return "untitled"
 }
 
-// mainArea 渲染中部区域，按终端宽度决定 Inspector 右侧或纵向压缩显示。
-func (a *App) mainArea() string {
-	streamView := a.agentStream.View()
-	if !a.state.Layout.ShowInspector {
-		return streamView
-	}
-	inspectorView := a.softInspector.View()
-	if a.state.Layout.Width >= inspectorWideMin {
-		return lipgloss.JoinHorizontal(lipgloss.Top, streamView, "  ", inspectorView)
-	}
-	return lipgloss.JoinVertical(lipgloss.Left, streamView, "", a.separatorLine(), inspectorView)
-}
-
-// separatorLine 渲染单条细线，用于区分主要区域而不使用边框。
-func (a *App) separatorLine() string {
-	width := a.state.Layout.Width
-	if width <= 0 {
-		width = 48
-	}
-	return theme.SubtleStyle().Render(strings.Repeat("─", width))
-}
-
-// fitViewToTerminal 将视图约束到当前终端尺寸，避免 resize 后自动换行或旧行残留。
-func (a *App) fitViewToTerminal(view string) string {
-	width := a.state.Layout.Width
-	height := a.state.Layout.Height
-	if width <= 0 {
-		return view
-	}
-	lines := strings.Split(view, "\n")
-	for i, line := range lines {
-		lines[i] = fitLine(line, width)
-	}
-	if height > 0 {
-		switch {
-		case len(lines) > height:
-			lines = lines[:height]
-		case len(lines) < height:
-			for len(lines) < height {
-				lines = append(lines, strings.Repeat(" ", width-1))
-			}
-		}
-	}
-	return strings.Join(lines, "\n")
-}
-
-// fitLine 截断并补齐单行显示宽度，保留 ANSI 样式同时防止终端自动 wrap。
-func fitLine(line string, width int) string {
-	if width <= 0 {
-		return line
-	}
-	target := width - 1
-	if target <= 0 {
-		return ""
-	}
-	fitted := theme.Truncate(line, target)
-	lineWidth := theme.DisplayWidth(fitted)
-	if lineWidth < target {
-		fitted += strings.Repeat(" ", target-lineWidth)
-	}
-	return fitted
-}
-
-// applyInitialLoaded 将 Gateway 初始 RPC 结果写入 ViewState。
-func (a *App) applyInitialLoaded(msg initialLoadedMsg) {
-	a.lastErr = msg.errText
-	a.state.Gateway.Connected = msg.connected
-	a.state.Gateway.Sessions = append([]gateway.SessionSummary(nil), msg.sessions...)
-	a.state.Gateway.Models = append([]gateway.ModelInfo(nil), msg.models...)
-	a.state.Gateway.ActiveModel = msg.activeModel
-	a.eventCh = msg.eventCh
-	if msg.errText != "" {
-		a.state.Runtime.Phase = state.RuntimePhaseError
-	}
-	if len(msg.sessions) > 0 {
-		active := msg.sessions[0]
-		a.state.Gateway.ActiveSess = &active
-	}
-	if msg.detail != nil {
-		a.state.Runtime.Tokens = state.TokenUsage{
-			Input:  msg.detail.Usage.Input,
-			Output: msg.detail.Usage.Output,
-			Total:  msg.detail.Usage.Total,
-		}
-		for _, item := range msg.detail.Stream {
-			a.appendStream(streamEntryFromItem(item))
-		}
-	}
-}
-
-// appendStream 以追加新 entry 的方式维护不可变 StreamEntry 序列。
 func (a *App) appendStream(entry state.StreamEntry) {
 	a.state.Stream = append(a.state.Stream, entry)
 }
@@ -1154,37 +765,6 @@ func (a *App) triggerCompact() tea.Cmd {
 	return nil
 }
 
-// debugLine 渲染调试模式下的最小运行信息。
-func (a *App) debugLine() string {
-	size := defaultTerminal
-	if a.state.Layout.Width > 0 || a.state.Layout.Height > 0 {
-		size = fmt.Sprintf("%dx%d", a.state.Layout.Width, a.state.Layout.Height)
-	}
-	return fmt.Sprintf(
-		"[debug] mode:%s  scenario:%s  events:%d  size:%s",
-		inputModeName(a.state.Mode),
-		a.scenario,
-		len(a.state.Stream),
-		size,
-	)
-}
-
-// streamEntryFromItem 将会话历史 DTO 映射为不可变 StreamEntry。
-func streamEntryFromItem(item gateway.StreamItem) state.StreamEntry {
-	return state.StreamEntry{
-		ID:        item.ID,
-		Type:      item.Kind,
-		Timestamp: item.CreatedAt,
-		Content:   item.Text,
-		Metadata: map[string]any{
-			"role":   item.Role,
-			"status": item.Status,
-			"done":   true,
-		},
-	}
-}
-
-// inputModeName 将输入模式转换为占位视图中的稳定文本。
 func inputModeName(mode state.InputMode) string {
 	switch mode {
 	case state.NormalMode:
@@ -1202,204 +782,4 @@ func emptyDash(value string) string {
 		return "-"
 	}
 	return value
-}
-
-type initialLoadedMsg struct {
-	connected   bool
-	sessions    []gateway.SessionSummary
-	detail      *gateway.SessionDetail
-	models      []gateway.ModelInfo
-	activeModel string
-	eventCh     <-chan gateway.GatewayEvent
-	errText     string
-}
-
-type gatewayEventMsg struct {
-	event  gateway.GatewayEvent
-	closed bool
-}
-
-// loadInitialCmd 通过 Gateway 客户端加载初始状态，并建立首个会话的事件订阅。
-func loadInitialCmd(client gateway.Client) tea.Cmd {
-	return func() tea.Msg {
-		ctx := context.Background()
-		msg := initialLoadedMsg{}
-		if _, err := client.Health(ctx); err != nil {
-			msg.errText = err.Error()
-			return msg
-		}
-		msg.connected = true
-		sessions, err := client.ListSessions(ctx)
-		if err != nil {
-			msg.errText = err.Error()
-			return msg
-		}
-		msg.sessions = sessions
-		models, err := client.ListModels(ctx)
-		if err != nil {
-			msg.errText = err.Error()
-			return msg
-		}
-		msg.models = models
-		if len(sessions) == 0 {
-			return msg
-		}
-		activeModel, err := client.GetModel(ctx, sessions[0].ID)
-		if err != nil {
-			msg.errText = err.Error()
-			return msg
-		}
-		msg.activeModel = activeModel
-		detail, err := client.LoadSession(ctx, sessions[0].ID)
-		if err != nil {
-			msg.errText = err.Error()
-			return msg
-		}
-		msg.detail = detail
-		eventCh, err := client.SubscribeEvents(ctx, sessions[0].ID)
-		if err != nil {
-			msg.errText = err.Error()
-			return msg
-		}
-		msg.eventCh = eventCh
-		return msg
-	}
-}
-
-// waitEventCmd 等待 Gateway 事件 channel 的下一条事件，保持异步事件逐条进入 Update。
-func waitEventCmd(events <-chan gateway.GatewayEvent) tea.Cmd {
-	return func() tea.Msg {
-		event, ok := <-events
-		return gatewayEventMsg{event: event, closed: !ok}
-	}
-}
-
-// submitMessageCmd 调用 GatewayClient 发送用户消息，并把 ACK 转成 reducer 可消费事件。
-func submitMessageCmd(client gateway.Client, sessionID string, text string) tea.Cmd {
-	return func() tea.Msg {
-		ack, err := client.SendMessage(context.Background(), sessionID, text)
-		if err != nil {
-			return gatewayEventMsg{event: errorEvent(err)}
-		}
-		return gatewayEventMsg{event: gateway.GatewayEvent{
-			Type:      gateway.EventRunStarted,
-			SessionID: ack.SessionID,
-			RunID:     ack.RunID,
-			Payload:   map[string]any{"message": ack.Message, "accepted": ack.Accepted},
-			At:        time.Now(),
-		}}
-	}
-}
-
-// resolvePermissionCmd 调用 GatewayClient 提交权限决策，并把完成结果转成 GatewayEvent。
-func resolvePermissionCmd(client gateway.Client, decision gateway.PermissionDecision) tea.Cmd {
-	return func() tea.Msg {
-		if err := client.ResolvePermission(context.Background(), decision); err != nil {
-			return gatewayEventMsg{event: errorEvent(err)}
-		}
-		text := "permission denied"
-		if decision.Allow {
-			text = "permission allowed"
-		}
-		return gatewayEventMsg{event: gateway.GatewayEvent{
-			Type:      gateway.EventPermissionResolved,
-			SessionID: decision.SessionID,
-			RunID:     decision.RunID,
-			Payload:   map[string]any{"decision": decision.Reason, "message": text},
-			At:        time.Now(),
-		}}
-	}
-}
-
-// answerQuestionCmd 调用 GatewayClient 提交 ask_user 回答，并把完成结果转成 GatewayEvent。
-func answerQuestionCmd(client gateway.Client, answer gateway.UserQuestionAnswer) tea.Cmd {
-	return func() tea.Msg {
-		if err := client.AnswerUserQuestion(context.Background(), answer); err != nil {
-			return gatewayEventMsg{event: errorEvent(err)}
-		}
-		return gatewayEventMsg{event: gateway.GatewayEvent{
-			Type:      gateway.EventUserQuestionAnswered,
-			SessionID: answer.SessionID,
-			RunID:     answer.RunID,
-			Payload:   map[string]any{"answer": answer.Text, "message": "answer submitted"},
-			At:        time.Now(),
-		}}
-	}
-}
-
-// errorEvent 将 GatewayClient RPC 错误包装成统一错误事件。
-func errorEvent(err error) gateway.GatewayEvent {
-	return gateway.GatewayEvent{
-		Type:    gateway.EventError,
-		Payload: map[string]any{"message": err.Error()},
-		At:      time.Now(),
-	}
-}
-
-// cancelRunCmd 调用 GatewayClient 取消运行中的 Agent，并把完成结果转成 GatewayEvent。
-func cancelRunCmd(client gateway.Client, sessionID string, runID string) tea.Cmd {
-	return func() tea.Msg {
-		if err := client.CancelRun(context.Background(), sessionID, runID); err != nil {
-			return gatewayEventMsg{event: errorEvent(err)}
-		}
-		return gatewayEventMsg{event: gateway.GatewayEvent{
-			Type:      gateway.EventRunCancelled,
-			SessionID: sessionID,
-			RunID:     runID,
-			Payload:   map[string]any{"message": "run cancelled by user"},
-			At:        time.Now(),
-		}}
-	}
-}
-
-// loadSessionCmd 切换到指定会话并建立新的事件订阅。
-func loadSessionCmd(client gateway.Client, sessionID string) tea.Cmd {
-	return func() tea.Msg {
-		detail, err := client.LoadSession(context.Background(), sessionID)
-		if err != nil {
-			return gatewayEventMsg{event: errorEvent(err)}
-		}
-		eventCh, err := client.SubscribeEvents(context.Background(), sessionID)
-		if err != nil {
-			return gatewayEventMsg{event: errorEvent(err)}
-		}
-		return sessionSwitchedMsg{sessionID: sessionID, detail: detail, eventCh: eventCh}
-	}
-}
-
-// deleteSessionCmd 调用 GatewayClient 删除会话。
-func deleteSessionCmd(client gateway.Client, sessionID string) tea.Cmd {
-	return func() tea.Msg {
-		// Gateway Client 接口暂无 DeleteSession，此处预留
-		return gatewayEventMsg{event: gateway.GatewayEvent{
-			Type:      gateway.EventSessionDeleted,
-			SessionID: sessionID,
-			Payload:   map[string]any{"id": sessionID, "message": "session deleted"},
-			At:        time.Now(),
-		}}
-	}
-}
-
-// sessionSwitchedMsg 表示会话切换完成。
-type sessionSwitchedMsg struct {
-	sessionID string
-	detail    *gateway.SessionDetail
-	eventCh   <-chan gateway.GatewayEvent
-}
-
-// sessionCreatedMsg 表示新会话创建完成。
-type sessionCreatedMsg struct {
-	Session *gateway.SessionSummary
-	err     error
-}
-
-// createSessionCmd 通过 GatewayClient 创建新会话。
-func createSessionCmd(client gateway.Client) tea.Cmd {
-	return func() tea.Msg {
-		summary, err := client.CreateSession(context.Background())
-		if err != nil {
-			return sessionCreatedMsg{err: err}
-		}
-		return sessionCreatedMsg{Session: summary}
-	}
 }
